move _linear_regression out
